:sparkles: Criando um hook customizado no contexto de carrinho onde vai ficar a função de add item no carrinho - mantendo ao maximo o padrão SOLID
