+++ src/common/context/Carrinho.js
@@ -1,3 +1,4 @@
+import { useContext } from 'react';
 import { useState } from 'react';
 import { createContext } from 'react';
 
@@ -11,4 +12,30 @@ export const CarrinhoProvider = ({ children }) => {
       {children}
     </CarrinhoContext.Provider>
   );
+};
+
+export const useCarrinhoContext = () => {
+  const { carrinho, setCarrinho } = useContext(CarrinhoContext);
+
+  const addProduto = (produto) => {
+    const temNoCarrinho = carrinho.some(
+      (itemCarrinho) => itemCarrinho.id === produto.id
+    );
+    if (!temNoCarrinho) {
+      produto.quantidade = 1;
+      return setCarrinho(() => [...carrinho, produto]);
+    }
+    setCarrinho(() =>
+      carrinho.map((itemCarrinho) => {
+        if (itemCarrinho.id === produto.id) itemCarrinho.quantidade += 1;
+        return itemCarrinho;
+      })
+    );
+  };
+
+  return {
+    carrinho,
+    setCarrinho,
+    addProduto,
+  };
 };
